去掉端口号  优化

+++ XSHCar/FunctionVC/Home/4SService/FourSServiceViewController.m
@@ -225,7 +225,7 @@
     {
         return;
     }
-    lastSelectedIndex = segmentView.selectedSegmentIndex;
+    lastSelectedIndex = (int)segmentView.selectedSegmentIndex;
     currentPage = 1;
     self.dataArray = nil;
     [self.table reloadData];
